requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
beautifulsoup4
python-dotenv